+++ packages/virtual-pad/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@wwawing/virtual-pad",
+  "version": "0.0.0",
+  "description": "WWA Wing Virtual Pad for smaprtphone",
+  "main": "src/VirtualPad.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "WWA Wing Team",
+  "license": "MIT"
+}
