+++ packages/virtual-pad/src/RotationControl.ts
@@ -1,0 +1,6 @@
+/**
+ * 画面が回転すると配置が変わるようにするJavaScriptファイルです。
+ */
+export default class RotationControl {
+
+}
